feat(order): implement order number generation logic
