+++ .gitignore
@@ -23,5 +23,5 @@ yarn-debug.log*
 yarn-error.log*
 **/.DS_Store
 
-
+data
